% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.chester_geo}
\alias{autoplot.chester_geo}
\title{Autoplot methods for chester_geo objects}
\usage{
\method{autoplot}{chester_geo}(object, ...)
}
\arguments{
\item{object}{An object of class chester_geo}

\item{...}{Additional parameters}
}
\value{

}
\description{
Autoplot methods for chester_geo objects
}
