% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_layers}
\alias{available_layers}
\title{Available Layers}
\format{
A character vector where each element is a layer that can be accessed through the API.
}
\usage{
available_layers
}
\description{
This data is a character vector containing the layers available to extract from the Chesterfield County OpenGeospace API. These layers can be passed as arguments into the \code{\link{get_geo_data()}} function
}
\keyword{datasets}
