% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chestergeo.R
\docType{package}
\name{chestergeo-package}
\alias{chestergeo}
\alias{chestergeo-package}
\title{chestergeo: Tools for Accessing Chesterfield County VA Geographical Data}
\description{
chestergeo provides tools for accessing publicly-available geographical data regarding Chesterfield County and Chesterfield County Public Schools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ccps-research-eval/chestergeo}
  \item \url{https://ccps-research-eval.github.io/chestergeo/}
  \item Report bugs at \url{https://github.com/ccps-research-eval/chestergeo/issues}
}

}
\author{
\strong{Maintainer}: Eric Ekholm \email{eric.ekholm@gmail.com}

}
\keyword{internal}
