% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_addresses.R
\name{get_chesterfield_addresses}
\alias{get_chesterfield_addresses}
\title{Get Chesterfield Addresses}
\usage{
get_chesterfield_addresses()
}
\value{
A 75-column dataframe and \code{sf} object containing various pieces of information about each address in the county, including the geographical coordinates, address, zoned schools, and more
}
\description{
Fetches all of the addresses in Chesterfield County from the county address service. This is a light wrapper around \code{\link{get_geo_data}}
}
\examples{
\dontrun{
#this will take a while to run
adrs <- get_chesterfield_addresses()
}
}
