% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_data.R
\name{get_geo_data}
\alias{get_geo_data}
\title{Get Geography Data}
\usage{
get_geo_data(layer)
}
\arguments{
\item{layer}{character. The name of the layer to retrieve data from. See \code{\link{layers_crosswalk}} for possible layers}
}
\value{
An \code{sf} object with information about the given layer, including a geometry column
}
\description{
This function retrieves data for the requested layer from the Chesterfield County Open Geospace API
}
\examples{
\dontrun{
ms_bounds <- get_geo_data(layer = "MiddleSchoolBoundary")
}
}
