% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_school_boundaries.R
\name{get_school_boundaries}
\alias{get_school_boundaries}
\title{Get School Boundaries}
\usage{
get_school_boundaries(level = c("elem", "middle", "high"))
}
\arguments{
\item{level}{Either 'elem', 'middle', or 'high'}
}
\value{
an \code{sf} object with the school boundaries for the chosen school level
}
\description{
This function allows you to retrieve the school boundaries for a given level. Note that it is a wrapper around the \code{\link{get_geo_data}} function
}
\examples{
\dontrun{
a <- get_school_boundaries("elem")
}
}
