% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layers_crosswalk}
\alias{layers_crosswalk}
\title{Layers Crosswalk}
\format{
A tibble with 3 columns
\describe{
\item{service}{name of the service}
\item{layer_num}{number of the layer within the service}
\item{layer}{name of the layer}
}
}
\usage{
layers_crosswalk
}
\description{
Lookup table for services and layers in the Chesterfield OpenGeospace. Note that layers are nested within services, and this table is mostly useful to navigate making requests to the API
}
\keyword{datasets}
